+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/e2ei/E2EIRepository.kt
@@ -77,8 +77,6 @@ interface E2EIRepository {
     suspend fun getOAuthRefreshToken(): Either<CoreFailure, String?>
     suspend fun nukeE2EIClient()
     suspend fun fetchFederationCertificates(): Either<CoreFailure, Unit>
-    suspend fun getCurrentClientCrlUrl(): Either<CoreFailure, String>
-    suspend fun getClientDomainCRL(url: String): Either<CoreFailure, ByteArray>
 }
 
 @Suppress("LongParameterList")
@@ -261,14 +259,4 @@ class E2EIRepositoryImpl(
     override suspend fun nukeE2EIClient() {
         e2EIClientProvider.nuke()
     }
-
-    override suspend fun getCurrentClientCrlUrl(): Either<CoreFailure, String> =
-        userConfigRepository.getE2EISettings().map {
-            (Url(it.discoverUrl).protocolWithAuthority)
-        }
-
-    override suspend fun getClientDomainCRL(url: String): Either<CoreFailure, ByteArray> =
-        wrapApiRequest {
-            acmeApi.getClientDomainCRL(url)
-         }
 }
